requirements.txt
amqp==5.1.1
anyascii==0.3.1
arrow==1.2.2
async-timeout==4.0.2
Babel==2.10.3
beautifulsoup4==4.7.1
bibtexparser==1.1.0
billiard==3.6.4.0
binaryornot==0.4.4
blinker==1.5
celery==5.1.2
celery-once==3.0.1
certifi==2022.6.15
chardet==5.0.0
charset-normalizer==2.1.1
click==7.1.2
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cookiecutter==1.7.3
Cython==0.29.32
Deprecated==1.2.13
dnspython==2.2.1
email-validator==1.1.3
et-xmlfile==1.1.0
eventlet==0.33.0
filedepot==0.8.0
Flask==1.1.4
Flask-BabelEx==0.9.4
Flask-Caching==1.10.1
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Principal==0.4.0
Flask-Script==2.0.6
Flask-Security==3.0.0
Flask-WTF==0.14.3
future==0.18.2
greenlet==1.1.3
gunicorn==20.1.0
html5lib==1.1
idna==3.3
ijson==2.4
importlib-metadata==4.12.0
isodate==0.6.1
itsdangerous==1.1.0
Jinja2==2.11.3
jinja2-time==0.2.0
joblib==1.1.0
kombu==5.2.4
lxml==4.9.1
Markdown==3.4.1
MarkupSafe==2.0.1
nltk==3.6.5
nose==1.3.7
numpy==1.23.2
openpyxl==3.0.10
oxrdflib==0.3.1
packaging==21.3
pandas==1.4.3
passlib==1.7.4
pip==20.0.2
pkg-resources==0.0.0
poyo==0.5.0
prompt-toolkit==3.0.30
psutil==5.9.1
puremagic==1.14
PyJWT==2.4.0
pyoxigraph==0.3.6
pyparsing==3.0.9
pyshp==2.3.1
python-dateutil==2.8.2
python-slugify==6.1.2
pytidylib==0.3.2
pytz==2022.2.1
rdflib==6.0.2
rdflib-jsonld==0.6.2
redis==4.3.4
redislite==6.2.805324
regex==2022.8.17
requests==2.28.1
requests-testadapter==0.3.0
sadi==0.5.7
scipy==1.9.1
sdd2rdf==1.0.3
setlr==0.2.18
setuptools==44.0.0
six==1.16.0
soupsieve==2.3.2.post1
speaklater==1.3
text-unidecode==1.3
toposort==1.7
tqdm==4.64.0
urllib3==1.26.12
vine==5.0.0
wcwidth==0.2.5
webencodings==0.5.1
WebOb==1.8.7
Werkzeug==1.0.1
wheel==0.37.1
whyis==2.0.0
wrapt==1.14.1
WTForms==3.0.1
xlrd==2.0.1
zipp==3.8.1

mod-wsgi==4.9.0
